pyproject.toml
[project]
name = "allele-discovery"
description = "A Snakemake workflow for discovering novel alleles with PacBio HiFi reads"
authors = [
    { name = "David H. O'Connor", email = "dhoconno@wisc.edu" },
    { name = "Nicholas R. Minor", email = "nrminor@wisc.edu" },
]
dependencies = [
    "biopython>=1.83",
    "loguru>=0.7.2",
    "pandas>=2.0.3",
    "pysam>=0.22.1",
    "snakemake>=7.32.4,<9",
    "pulp==2.7.0",
]
requires-python = ">= 3.8"
version = "0.1.0"

[build-system]
requires = ["setuptools >= 40.8.0", "cffi >=1.15.0,!=1.17.0"]
build-backend = "setuptools.build_meta:__legacy__"

[tool.setuptools]
py-modules = ["scripts"]

[tool.pixi.pypi-dependencies]
allele-discovery = { path = ".", editable = true }

# NOTE:
# --------------------------------------------------------------------------------------
# as of 2024-09-12, exonerate and clustalo do not support ARM/Apple Silicon Macs. As
# such, this pipeline does not yet support ARM macs.
# --------------------------------------------------------------------------------------

[tool.pixi.project]
channels = ["conda-forge", "bioconda", "anaconda"]
platforms = ["osx-64", "linux-64", "linux-aarch64"]

[tool.pixi.dependencies]
pbaa = ">=1.0.3,<2"
samtools = ">=1.20,<2"
bbmap = ">=39.8,<40"
exonerate = ">=2.4.0,<3"
clustalo = ">=1.2.4,<2"
minimap2 = ">=2.1.1,<3"
gffread = ">=0.9.12,<0.13"
pigz = ">=2.8,<3"
unzip = ">=6.0,<7"
picard = ">=2.20.4,<4"
muscle = ">=5.1,<6"
vsearch = ">=2.28.1,<3"
openpyxl = ">=3.1.5,<4"
pyfasta = ">=0.5.2,<0.6"
perl = ">=5.32.1,<6"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipython>=8.12.3",
    "jupyter>=0.6.4",
    "mamba>=0.11.3",
    "marimo[sql]>=0.8.14",
    "polars>=1.7.0",
    "pytest>=8.3.3",
    "ruff",
    "tox>=4.18.1",
]
compile-bytecode = true

[project.scripts]
"alleled" = "bin.__main__:main"
